Use graph._c_graph instead of global USE_C_API
Looking at graph._c_graph should be safer. The logic would be as
follows. You need to set _USE_C_API global when you create the graph.
Once the graph is created, it is either in the "C API mode" or in the
old mode. All ops created on this graph inherit its mode and don't look
at the global.

PiperOrigin-RevId: 161620767