CPU IR Emitter: Lower slice operations with memcpy.
This changes the way the CPU backend emits slice operations
so that the maximal contiguous blocks get copied with memcpy.

Without this change, LLVM's optimizers prefer to unroll the
elementwise loops emitted for doing slice. This leads to increased
code size. With this change, the inliner only inlines the memcpy
when it is small enough.

This change also adds a test that exercises the logic to find
maximal contiguous blocks.

PiperOrigin-RevId: 162249449