Variant objects can now be used inside the TensorFlow execution graph.
Includes modifications and bugfixes to Variant and Tensor objects.

The most significant change is that a Variant storing a
VariantTensorDataProto acts like the data type inside the stored proto.

This means that its TypeName() is the typename of the pre-serialized data
object, and .get<> has been renamed to MaybeDecodeAndGet<>.  Calling
MaybeDecodeAndGet<OriginalType> performs a Decode under the hood.
Calling MaybeDecodeAndGet<OriginalType> on a **const** Variant of
VariantTensorDataProto causes a CHECK failure.

Also added proper TypeName() for unboxed primitive types (int, ...).

PiperOrigin-RevId: 165510197